Rename: 수정 이미지파일 경로
